requirements.txt
aiogram==3.3.0
environs==9.5.0
pytz==2023.3.post1
python-dateutil==2.8.2